Bundle AssistantDetails in ConversationLayout
